refactor some code
