+++ src/css/article.module.scss
@@ -75,7 +75,8 @@
                 bottom: 0;
                 margin: auto;
                 width: auto !important;
-                height: 100% !important;
+                height: auto !important;
+                max-height: 100%;
             }
             .noImage {
                 position: absolute;
@@ -117,14 +118,16 @@
                 display: grid;
                 align-items: center;
                 overflow: hidden;
+                background-color: #333;
+                border-radius: 5px;
                 img {
                     position: absolute;
                     top: 0;
                     left: 0;
                     right: 0;
                     bottom: 0;
-                    width: 100%;
-                    height: auto;
+                    width: auto !important;
+                    height: 100% !important;
                     margin: auto;
                 }
                 &:not(:nth-of-type(1)) {
