+++ src/css/blog.module.scss
@@ -27,8 +27,9 @@
             border-radius: 10px;
             img {
                 position: absolute;
-                width: 100% !important;
+                width: auto !important;
                 height: auto !important;
+                max-height: 100%;
                 transition: transform 0.5s;
                 top: 0;
                 left: 0;
