+++ src/css/layout.module.scss
@@ -3,7 +3,7 @@
 
 .html {
     overflow-x: hidden;
-    background-color: #1f1f1f;
+    background-color: #242526;
     color: white;
     font-family: "Open Sans", "Noto Sans JP", sans-serif;
     font-size: 10px;
@@ -20,7 +20,7 @@
         width: 95%;
         margin: auto;
         padding-top: 10px;
-        background-color: #1f1f1f;
+        background-color: #242526;
     }
 }
