+++ src/css/page.module.scss
@@ -318,7 +318,7 @@
             min-width: 550px;
             aspect-ratio: 3 / 2;
             border-radius: 10px;
-            background-color: #555;
+            background-color: #333;
             display: grid;
             align-items: center;
             &:not(:nth-of-type(1)) {
